Refactor find-python3.sh usage in create-serverless.sh
It was never correct to use "|| return" when sourcing find-python3.sh since we're not in a function context. Using "|| exit 1" achieves the desired effect, and is also more concise than the previous if statement.

Source find-python3.sh using a path relative to create-serverless.sh, as we already do for invoking get-instance.sh. While we could use ${BASH_SOURCE[0]} instead of $0, there is no need to do so as create-serverless.sh is never sourced.